figma-transformer: extract transform diagnostics builder
